Fix presenters
