+++ src/domain/pnl/asset.go
@@ -12,12 +12,25 @@ package pnl
 // 	UDST Symbol = "UDST"
 // )
 
+const (
+	Crypto AssetType = "Crypto"
+	Forex  AssetType = "Forex"
+	Stock  AssetType = "Stock"
+)
+
+type AssetType string
+
+func (a *AssetType) getValidTypes() []AssetType {
+	return []AssetType{Crypto, Forex, Stock}
+}
+
 // Asset represents individual assets like BTC, USD, EUR, etc. the Symbol property uniquely identifies an asset.
 type Asset struct {
 	Symbol      string
 	Color       string
 	Name        string
 	CountryCode string
+	Type        AssetType
 }
 
 // Assets is the repository that handles the CRUD of Assets
