+++ src/domain/pnl/assets.go
@@ -3,16 +3,19 @@ package pnl
 import (
 	"errors"
 	"regexp"
+	"slices"
 )
-
-const HEXCOLORPATTERN = `^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$`
 
 type InvalidAsset error
 
 func (a *Asset) Validate() (*Asset, error) {
-	re := regexp.MustCompile(HEXCOLORPATTERN)
+	hexColorPattern := `^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$`
+	re := regexp.MustCompile(hexColorPattern)
 	if !re.MatchString(a.Color) {
 		return nil, InvalidAsset(errors.New("Wrong color string"))
+	}
+	if slices.Contains(a.Type.getValidTypes(), a.Type) {
+		return nil, InvalidAsset(errors.New("Invalid Asset Symbol"))
 	}
 	if len(a.Symbol) < 3 || len(a.Symbol) > 8 {
 		return nil, InvalidAsset(errors.New("Invalid Asset Symbol"))
