--- src/presenters/uiHandlers.go
@@ -1,1 +1,0 @@
-package presenters
