+++ src/rest/pnlHandlers.go
@@ -1,0 +1,13 @@
+package presenters
+
+import (
+	"controtto/src/app/querying"
+
+	"github.com/gofiber/fiber/v2"
+)
+
+func avgBuyPrice(tpq querying.TradingPairsQuerier) func(*fiber.Ctx) error {
+	return func(c *fiber.Ctx) error {
+		return nil
+	}
+}
