+++ src/rest/routes.go
@@ -14,41 +14,35 @@ func Run(port string, m *managing.Service, q *querying.Service) {
 	engine.Debug(true)
 	app := fiber.New(fiber.Config{
 		Views: engine,
-		// ViewsLayout: "layouts/main",
 		ErrorHandler: func(c *fiber.Ctx, err error) error {
 			return c.Status(fiber.StatusInternalServerError).SendString(err.Error())
 		},
 	})
+	app.Static("/assets", "./public/assets")
 
-	// UI
-	app.Get("/healthcheck/price", checkPrice(q.MarketQuerier))
-	app.Get("/ui/pairs/form", newPairForm(q.AssetQuerier))
-	app.Get("/ui/pairs/table", pairsTable(q.TradingPairQuerier))
-	app.Get("/ui/pairs/:id/transactions/table", transactionTable(q.TradingPairQuerier))
-	app.Get("/ui/pairs/:id/cards", pairCards(q.TradingPairQuerier))
-	app.Get("/ui/pairs/:id/transactions/form", newTransactionForm(q.TradingPairQuerier))
-	app.Get("/ui/assets/form", newAssetForm)
-	app.Get("/pairs/:id/", pairSection)
-	app.Get("/pairs/:id/transactions/export", transactionExport(q.TradingPairQuerier))
-	app.Get("/pairs/", pairsSection)
-	app.Get("/dashboard", dashboardSection(q.TradingPairQuerier))
 	// GET
 	app.Get("/", Home)
-	app.Get("/tables", Tables)
-	// app.Get("/pairs/AvgBuyPrice/:id", avgBuyPrice(q.TradingPairQuerier))
-	app.Get("/users", TotalUsers)
+	app.Get("/pairs/", pairsSection)
+	app.Get("/pairs/:id/", pairSection)
+	app.Get("/ui/assets/form", newAssetForm)
+	app.Get("/ui/pairs/form", newPairForm(q.AssetQuerier))
+	app.Get("/healthcheck/price", checkPrice(q.MarketQuerier))
+	app.Get("/ui/pairs/table", pairsTable(q.TradingPairQuerier))
+	app.Get("/dashboard", dashboardSection(q.TradingPairQuerier))
+	app.Get("/ui/pairs/:id/cards", pairCards(q.TradingPairQuerier))
+	app.Get("/pairs/AvgBuyPrice/:id", avgBuyPrice(q.TradingPairQuerier))
+	app.Get("/pairs/:id/transactions/export", transactionExport(q.TradingPairQuerier))
+	app.Get("/ui/pairs/:id/transactions/table", transactionTable(q.TradingPairQuerier))
+	app.Get("/ui/pairs/:id/transactions/form", newTransactionForm(q.TradingPairQuerier))
 	// DELETE
+	app.Delete("/empty", empty())
 	app.Delete("/pairs/:id", deleteTradingPair(m.TradingPairManager))
 	app.Delete("/transactions/:id", deleteTransaction(m.TradingPairManager))
 	// POST
-	app.Post("/pairs/:id/transactions", newTransaction(m.TradingPairManager))
-	app.Post("/pairs/:id/transactions/upload", newTransactionImport(m.TradingPairManager))
 	app.Post("/assets", newAsset(m.AssetCreator))
 	app.Post("/pairs", newTradingPair(m.TradingPairManager))
-	app.Delete("/empty", func(c *fiber.Ctx) error {
-		return c.SendString("")
-	})
+	app.Post("/pairs/:id/transactions", newTransaction(m.TradingPairManager))
+	app.Post("/pairs/:id/transactions/upload", newTransactionImport(m.TradingPairManager))
 
-	app.Static("/assets", "./public/assets")
 	log.Fatal(app.Listen("0.0.0.0" + ":" + port))
 }
