+++ src/rest/transactionHandlers.go
@@ -9,161 +9,11 @@ import (
 
 	"fmt"
 	"log/slog"
-	"math/rand"
 	"slices"
 	"time"
 
 	"github.com/gofiber/fiber/v2"
 )
-
-func dashboardSection(aq querying.TradingPairsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		slog.Info("Dashboard requested")
-		req := querying.ListTradingPairsReq{}
-		resp, err := aq.ListTradingPairs(req)
-		if err != nil {
-			return c.Render("toastErr", fiber.Map{
-				"Title": "Error",
-				"Msg":   err,
-			})
-		}
-		return c.Render("dashboardSection", fiber.Map{
-			"Title": "Trading Pairs",
-			"Pairs": resp.Pairs,
-		})
-	}
-
-}
-
-func pairsSection(c *fiber.Ctx) error {
-	slog.Info("Pairs Section")
-	return c.Render("pairsSection", fiber.Map{})
-}
-
-func pairSection(c *fiber.Ctx) error {
-	slog.Info("Pair Section")
-	return c.Render("pairSection", fiber.Map{
-		"PairID": c.Params("id"),
-	})
-}
-
-// Home hanlder reders the homescreen
-func Home(c *fiber.Ctx) error {
-	slog.Info("HOME")
-	// render index template
-	return c.Render("main", fiber.Map{})
-}
-
-func checkPrice(mq querying.MarketsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		base := c.Query("base")
-		quote := c.Query("quote")
-		req := querying.QueryMarketReq{
-			AssetSymbolA: base,
-			AssetSymbolB: quote,
-		}
-		resp, err := mq.GetMarketPrice(req)
-		if err != nil {
-			return c.SendString(err.Error())
-		}
-		return c.SendString(fmt.Sprintf("%f", resp.Price))
-	}
-}
-
-func newPairForm(aq querying.AssetsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		req := querying.QueryAssetsReq{}
-		resp, err := aq.ListAssets(req)
-		if err != nil {
-			return c.Render("toastErr", fiber.Map{
-				"Title": "Error",
-				"Msg":   err,
-			})
-		}
-		return c.Render("newPairForm", fiber.Map{
-			"Title":  "New Pair",
-			"Assets": resp.Assets,
-		})
-	}
-}
-
-func newAssetForm(c *fiber.Ctx) error {
-	slog.Info("Create Asset UI requested")
-	return c.Render("newAssetForm", fiber.Map{
-		"Title": "New Asset",
-	})
-}
-
-// Tables handler that renderizer the tables view and returns it to the client
-func pairsTable(aq querying.TradingPairsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		slog.Info("Pairs table requested")
-		req := querying.ListTradingPairsReq{}
-		resp, err := aq.ListTradingPairs(req)
-		if err != nil {
-			return c.Render("toastErr", fiber.Map{
-				"Title": "Error",
-				"Msg":   err,
-			})
-		}
-		return c.Render("pairsTable", fiber.Map{
-			"Title": "Trading Pairs",
-			"Pairs": resp.Pairs,
-		})
-	}
-}
-
-func avgBuyPrice(tpq querying.TradingPairsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		return nil
-	}
-}
-
-// marketPrice handler that reutnr a string witht the total amount of money.
-func marketPrice(aq querying.MarketsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-
-		id := c.Params("id")
-		slog.Info("Requesting market price", "market", id)
-		req := querying.QueryMarketReq{
-			AssetSymbolA: id,
-		}
-		resp, err := aq.GetMarketPrice(req)
-		if err != nil {
-			return c.SendString("--")
-		}
-		// render index template
-		return c.SendString(fmt.Sprintf("%.2f", resp.Price))
-	}
-}
-
-// TotalUsers handler that reutnr a string witht the total amount of users.
-func TotalUsers(c *fiber.Ctx) error {
-	slog.Info("USERS")
-	rand.New(rand.NewSource(time.Now().UnixNano()))
-	randomNumber := rand.Intn(101)
-	return c.SendString(fmt.Sprintf("%d", randomNumber))
-}
-
-// Tables handler that renderizer the tables view and returns it to the client
-func Tables(c *fiber.Ctx) error {
-	slog.Info("TABLES")
-	return c.Render("tables", fiber.Map{
-		"Title":    "Hello, World!",
-		"Selected": "tables",
-	})
-}
-
-func NewPair(c *fiber.Ctx) error {
-	slog.Info("Creating new asset")
-	slog.Info(string(c.Body()))
-	return c.Send(c.Body())
-}
-
-func Assets(c *fiber.Ctx) error {
-	slog.Info("Assets requested")
-	return nil
-}
 
 func deleteTransaction(tpm managing.TradingPairsManager) func(*fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
@@ -407,28 +257,6 @@ func newTransaction(tpm managing.TradingPairsManager) func(*fiber.Ctx) error {
 			"Extra": resp.RecordTime.Format("15h 04m 05s"),
 		})
 
-	}
-}
-
-func pairCards(tpq querying.TradingPairsQuerier) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		id := c.Params("id")
-		req := querying.GetTradingPairReq{
-			TPID:             id,
-			WithCalculations: true,
-		}
-		resp, err := tpq.GetTradingPair(req)
-		if err != nil {
-			return c.Render("toastErr", fiber.Map{
-				"Title": "Error",
-				"Msg":   err,
-			})
-		}
-		slices.Reverse(resp.Pair.Transactions)
-		return c.Render("pairCards", fiber.Map{
-			"Today": time.Now().Format("Mon Jan 02 15:04 2006"),
-			"Pair":  resp.Pair,
-		})
 	}
 }
